Change the author in setup.py
I believe that the author should either be the current author (those of
us in the Python LTI Initiative), or the original author. Since the
field has been changed before, I've opted to update it to the Python LTI
Initiative, but if that seems inappropriate than we should probably
revert it back to its original author, Anson MacKeracher.