Do not require secret for tool provider
Actually, we want to prefer _not_ having the secret passed in to the
provider if at all possible, because it encourages shortcuts when
creating the validator that can easily cause timing attacks. Plus, the
secret is only really useful to methods dealing with outcomes, so for
the general case it's not needed anyway.